Put the getTraceabilityLogs function in DAO mother class
it was defined in all subclasses so it is better to define it only once in the mother class
