+++ phis2-ws/src/main/java/phis2ws/service/dao/manager/DAO.java
@@ -10,6 +10,7 @@ package phis2ws.service.dao.manager;
 import java.util.List;
 import opensilex.service.dao.exception.DAODataErrorAggregateException;
 import opensilex.service.dao.exception.ResourceAccessDeniedException;
+import phis2ws.service.model.User;
 
 /**
  * DAO mother class.
@@ -17,6 +18,12 @@ import opensilex.service.dao.exception.ResourceAccessDeniedException;
  * @param <T> the type of object handled
  */
 public abstract class DAO<T> {
+    
+    /**
+     * Remote user address for logging
+     */
+    public String remoteUserAdress;
+    public User user;
 
     /**
      * Creates in the storage the list of objects given.
@@ -77,5 +84,20 @@ public abstract class DAO<T> {
             throws DAODataErrorAggregateException, ResourceAccessDeniedException, Exception {
         checkBeforeCreation(annotations);
         return create(annotations);
+    }
+    
+     /**
+     * @return the traceability logs
+     */
+    protected String getTraceabilityLogs() {
+        String log = "";
+        if (remoteUserAdress != null) {
+            log += "IP Address " + remoteUserAdress + " - ";
+        }
+        if (user != null) {
+            log += "User : " + user.getEmail() + " - ";
+        }
+        
+        return log;
     }
 }
